finished tests

+++ lab-carlo/controller/car-controller.js
@@ -1,27 +1,54 @@
 'use strict';
-//
-// const Promise = require('bluebird');
-// const fs = Promise.promisifyAll(require('fs'), {suffix: 'Prom'});
-// const createError = require('http-errors');
+
+const Promise = require('bluebird');
+const createError = require('http-errors');
 const Car = require('../model/cars');
 
 module.exports = exports = {};
 
 
-exports.createCar = function(auto) {
-  return new Car(auto).save();
+exports.createCar = function(req, res, car) {
+  if(!car) return Promise.reject(createError(404, 'Car is required'));
+  new Car(req.body).save()
+  .then(car => {
+    res.json(car);
+  })
+  .catch(err => res.status(404).send(err.message));
 };
 
-exports.fetchCar = function(id) {
-  return Car.findById(id);
+exports.fetchCar = function(id, res) {
+  if(!id) return Promise.reject(createError(400, 'ID is required'));
+  return Car.findById(id)
+  .then(car => {
+    console.log(car);
+    res.json(car);
+  })
+  .catch(err => res.status(404).send(err.message));
 
 };
 
-exports.deleteCar = function(id){
-  return Car.findByIdAndRemove(id);
+exports.fetchAllCars = function(res) {
+  return Car.find()
+  .then(car => res.json(car))
+  .catch(err => res.status(400).send(err.message));
 };
 
-exports.putCar = function(auto, id) {
-  return Car.findByIdAndUpdate(id, auto, {new: true});
+exports.deleteCar = function(id, res){
+  if(!id) return Promise.reject(createError(400, 'ID is required'));
+
+  Car.findByIdAndRemove(id)
+  .then(car => {
+    console.log(car);
+    res.sendStatus(204);
+  })
+  .catch(err => res.send(err));
+};
+
+exports.putCar = function(req, res, id) {
+  if(!id) return Promise.reject(createError(400, 'ID is required'));
+
+  Car.findByIdAndUpdate(id, req.body, {new: true})
+  .then(car => res.json(car))
+  .catch(err => res.status(400).send(err.message));
 
 };
